basic processing (bad)

+++ diff_map.py
@@ -5,6 +5,8 @@ from pathlib import Path
 def generate_diff_maps(im1, im2, output_dir="outputs"):
     assert im1.shape == im2.shape
 
+    im1 = im1.astype(np.float32)
+    im2 = im2.astype(np.float32)
 # get mse for each sq
     mse_b = (im1[:,:,0] - im2[:,:,0]) ** 2
     mse_g = (im1[:,:,1] - im2[:,:,1]) ** 2
@@ -28,6 +30,6 @@ def generate_diff_maps(im1, im2, output_dir="outputs"):
     cv2.imwrite(output_dir/"mse_blue.png", sq_b_vis)
 
 if __name__ == "__main__":
-    night = cv2.imread("images/basic_ps.jpg").astype(np.float32)
-    day = cv2.imread("images/day.jpg").astype(np.float32)
+    night = cv2.imread("images/basic_ps.jpg")
+    day = cv2.imread("images/day.jpg")
     generate_diff_maps(night, day)
